requirements.txt
flask[async]
flask-compress
aiohttp
asyncio
Pillow
requests
PyPDF2
python-docx
tiktoken